+++ app/src/app/components/toolbar/toolbar.component.ts
@@ -1,8 +1,7 @@
 import { Component, OnInit, ViewEncapsulation, Output, EventEmitter, Input, ViewChild } from '@angular/core';
-import { ActivatedRoute, NavigationEnd, Router } from '@angular/router';
 import { PopoverContentComponent } from 'ngx-smart-popover';
 import { ValidationData } from 'src/app/classes/serializable';
-import { RestApiConnectorService } from 'src/app/services/connectors/rest-api/rest-api-connector.service';
+import { AuthenticationService } from 'src/app/services/connectors/authentication/authentication.service';
 import { EditorService } from 'src/app/services/editor/editor.service';
 import { SidebarService } from 'src/app/services/sidebar/sidebar.service';
 
@@ -22,9 +21,11 @@ export class ToolbarComponent implements OnInit {
     public validationData: ValidationData = null;
 
     public get editing(): boolean { return this.editorService.editing; }
-    public get editable(): boolean { return this.editorService.editable; } 
+    public get editable(): boolean { return this.editorService.editable; }
 
-    constructor(private sidebarService: SidebarService, private editorService: EditorService, private restAPIService: RestApiConnectorService) {}
+    public get isLoggedIn(): boolean { return this.authenticationService.isLoggedIn; }
+
+    constructor(private sidebarService: SidebarService, private editorService: EditorService, private authenticationService: AuthenticationService) {}
 
     ngOnInit() {}