+++ app/src/app/views/landing-page/landing-page.component.ts
@@ -1,7 +1,9 @@
-import { Component, OnInit, ViewEncapsulation } from '@angular/core';
+import { Component, OnDestroy, OnInit, ViewEncapsulation } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { Role } from 'src/app/classes/authn/role';
 import { ConfirmationDialogComponent } from 'src/app/components/confirmation-dialog/confirmation-dialog.component';
+import { AuthenticationService } from 'src/app/services/connectors/authentication/authentication.service';
 import { RestApiConnectorService } from 'src/app/services/connectors/rest-api/rest-api-connector.service';
 import { stixRoutes } from "../../app-routing-stix.module";
 
@@ -11,36 +13,60 @@ import { stixRoutes } from "../../app-routing-stix.module";
   styleUrls: ['./landing-page.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
-export class LandingPageComponent implements OnInit {
-    
+export class LandingPageComponent implements OnInit, OnDestroy {
+    private loginSubscription;
+    public pendingUsers;
     public routes: any[] = [];
-    constructor(private restApiConnector: RestApiConnectorService, private dialog: MatDialog, private router: Router) {
+
+    constructor(private restApiConnector: RestApiConnectorService, private authenticationService: AuthenticationService, private dialog: MatDialog, private router: Router) {
         this.routes = stixRoutes;
     }
 
+    public get isAdmin(): boolean { return this.authenticationService.isAuthorized([Role.ADMIN]); }
+    public get isLoggedIn(): boolean { return this.authenticationService.isLoggedIn; }
+
     ngOnInit() {
-        // bug the user about editing their organization identity
-        let subscription = this.restApiConnector.getOrganizationIdentity().subscribe({
-            next: (identity) => {
-                if (identity.name == "Placeholder Organization Identity") {
-                    let prompt = this.dialog.open(ConfirmationDialogComponent, {
-                        maxWidth: "35em",
-                        data: { 
-                            message: '### Your organization identity has not yet been set.\n\nYour organization identity is used for attribution of edits you make to objects in the knowledge base and is attached to published collections. Currently, a placeholder is being used.\n\nUpdate your organization identity now?',
-                            yes_suffix: "edit my identity now",
-                            no_suffix: "edit my identity later"
-                        }
-                    })
-                    let prompt_subscription = prompt.afterClosed().subscribe({
-                        next: (prompt_result) => {
-                            if (prompt_result) this.router.navigate(["/admin/org-identity"]);
-                        },
-                        complete: () => { prompt_subscription.unsubscribe(); }
-                    })
-                }
+        this.loginSubscription = this.authenticationService.onLogin.subscribe({
+            next: (event) => { this.openOrgIdentityDialog(); }
+        });
+        let userSubscription = this.restApiConnector.getAllUserAccounts({status: ["pending"]}).subscribe({
+            next: (results) => {
+                let users = results as any;
+                if (users && users.length) this.pendingUsers = users.length;
             },
-            complete: () => subscription.unsubscribe()
+            complete: () => userSubscription.unsubscribe()
         })
+        this.openOrgIdentityDialog();
     }
 
+    // bug the admin about editing their organization identity
+    private openOrgIdentityDialog(): void {
+        if (this.authenticationService.isAuthorized([Role.ADMIN])) {
+            let subscription = this.restApiConnector.getOrganizationIdentity().subscribe({
+                next: (identity) => {
+                    if (identity.name == "Placeholder Organization Identity") {
+                        let prompt = this.dialog.open(ConfirmationDialogComponent, {
+                            maxWidth: "35em",
+                            data: {
+                                message: '### Your organization identity has not yet been set.\n\nYour organization identity is used for attribution of edits you make to objects in the knowledge base and is attached to published collections. Currently, a placeholder is being used.\n\nUpdate your organization identity now?',
+                                yes_suffix: "edit my identity now",
+                                no_suffix: "edit my identity later"
+                            }
+                        })
+                        let prompt_subscription = prompt.afterClosed().subscribe({
+                            next: (prompt_result) => {
+                                if (prompt_result) this.router.navigate(["/admin/org-identity"]);
+                            },
+                            complete: () => { prompt_subscription.unsubscribe(); }
+                        })
+                    }
+                },
+                complete: () => subscription.unsubscribe()
+            })
+        }
+    }
+
+    ngOnDestroy() {
+        this.loginSubscription.unsubscribe();
+    }
 }