+++ app/src/app/views/stix/relationship/relationship-view/relationship-view.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { forkJoin } from 'rxjs';
 import { Relationship } from 'src/app/classes/stix/relationship';
 import { StixObject, stixTypeToAttackType } from 'src/app/classes/stix/stix-object';
+import { AuthenticationService } from 'src/app/services/connectors/authentication/authentication.service';
 import { RestApiConnectorService } from 'src/app/services/connectors/rest-api/rest-api-connector.service';
 import { StixViewPage } from '../../stix-view-page';
 
@@ -28,8 +29,8 @@ export class RelationshipViewComponent extends StixViewPage implements OnInit {
         })
     }
 
-    constructor(private restApiService: RestApiConnectorService) { 
-        super()
+    constructor(private restApiService: RestApiConnectorService, authenticationService: AuthenticationService) { 
+        super(authenticationService)
     }
 
     ngOnInit(): void {