+++ app/src/app/views/stix/software/software-list/software-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthenticationService } from 'src/app/services/connectors/authentication/authentication.service';
 
 @Component({
   selector: 'app-software-list',
@@ -6,8 +7,9 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./software-list.component.scss']
 })
 export class SoftwareListComponent implements OnInit {
+    public get canEdit(): boolean { return this.authenticationService.canEdit(); }
 
-    constructor() { }
+    constructor(private authenticationService: AuthenticationService) { }
 
     ngOnInit() {}