+++ app/src/app/views/stix/technique/technique-view/technique-view.component.ts
@@ -1,29 +1,25 @@
-import { ChangeDetectorRef, Component, OnInit, ViewEncapsulation } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import { AfterContentChecked, ChangeDetectorRef, Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { Technique } from 'src/app/classes/stix/technique';
+import { AuthenticationService } from 'src/app/services/connectors/authentication/authentication.service';
 import { StixViewPage } from '../../stix-view-page';
 import { RestApiConnectorService } from "src/app/services/connectors/rest-api/rest-api-connector.service";
 
 @Component({
-  selector: 'app-technique-view',
-  templateUrl: './technique-view.component.html',
-  styleUrls: ['./technique-view.component.scss'],
-  encapsulation: ViewEncapsulation.None
+    selector: 'app-technique-view',
+    templateUrl: './technique-view.component.html',
+    styleUrls: ['./technique-view.component.scss'],
+    encapsulation: ViewEncapsulation.None
 })
-export class TechniqueViewComponent extends StixViewPage implements OnInit {
+export class TechniqueViewComponent extends StixViewPage implements OnInit, AfterContentChecked {
+    public get technique(): Technique { return this.config.object as Technique; }
 
-    public get technique(): Technique { 
-        return this.config.object as Technique; 
-    }
-
-    constructor(private restApiConnector: RestApiConnectorService, private route: ActivatedRoute, private ref: ChangeDetectorRef) {
-        super();
+    constructor(private ref: ChangeDetectorRef, authenticationService: AuthenticationService, private restApiConnector: RestApiConnectorService) {
+        super(authenticationService);
     }
 
     ngOnInit() {
-        let technique = this.config.object as Technique;
-        if ( technique.firstInitialized ) {
-            technique.initializeWithDefaultMarkingDefinitions(this.restApiConnector);
+        if (this.technique.firstInitialized) {
+            this.technique.initializeWithDefaultMarkingDefinitions(this.restApiConnector);
         }
     }